\name{IBFelix-package}
\alias{IBFelix-package}
\alias{IBFelix}
\docType{package}
\title{
\packageTitle{IBFelix}
}
\description{
\packageDescription{IBFelix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IBFelix}
\packageIndices{IBFelix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{IBFelix}

Maintainer: \packageMaintainer{IBFelix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
